move parsetree sizing tests to its own class
