+++ src/test/java/com/github/giedomak/telepathdb/datamodels/parsetree/ParseTreeTest.kt
@@ -1,6 +1,5 @@
-package com.github.giedomak.telepathdb.datamodels
+package com.github.giedomak.telepathdb.datamodels.parsetree
 
-import com.github.giedomak.telepathdb.datamodels.parsetree.ParseTree
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import java.util.*
@@ -32,89 +31,6 @@ class ParseTreeTest {
         val actual = root.postOrderTreeWalk().map { it.id }.collect(Collectors.toList()).toString()
 
         // then
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun subtreesOfSize2() {
-        // Input:
-        //       CONCATENATION
-        //        /   |   \
-        //       a    b  UNION
-        //                / \
-        //               c   d
-        val child = ParseTreeTest.create1LevelParseTree(ParseTree.UNION, listOf("c", "d"), false)
-        val input = ParseTreeTest.create1LevelParseTree(ParseTree.CONCATENATION, listOf("a", "b"))
-        input.children.add(child)
-
-        // Expected:
-        //   CONCATENATION   UNION
-        //       /  \         / \
-        //      a    b       c   d
-        val expected = listOf(
-                ParseTreeTest.create1LevelParseTree(ParseTree.CONCATENATION, listOf("a", "b"), true),
-                ParseTreeTest.create1LevelParseTree(ParseTree.UNION, listOf("c", "d"), false)
-        )
-
-        val actual = input.subtreesOfSize(2)
-
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun subtreesOfSize2Complex() {
-        // Input:
-        //         CONCATENATION
-        //        /  |     |  |  \
-        //       a  UNION  e  f   g
-        //          / | \
-        //         b  c  d
-        val child = ParseTreeTest.create1LevelParseTree(ParseTree.UNION, listOf("b", "c", "d"), false)
-        val input = ParseTreeTest.create1LevelParseTree(ParseTree.CONCATENATION, listOf("a", "e", "f", "g"))
-        input.children.add(1, child)
-
-        // Expected:
-        //     UNION   UNION    CONCATENATION    CONCATENATION
-        //      / \     / \         /   \            /   \
-        //     b   c   c   d       e     f          f     g
-        val expected = listOf(
-                ParseTreeTest.create1LevelParseTree(ParseTree.UNION, listOf("b", "c"), false),
-                ParseTreeTest.create1LevelParseTree(ParseTree.UNION, listOf("c", "d"), false),
-                ParseTreeTest.create1LevelParseTree(ParseTree.CONCATENATION, listOf("e", "f"), true),
-                ParseTreeTest.create1LevelParseTree(ParseTree.CONCATENATION, listOf("f", "g"), true)
-        )
-
-        val actual = input.subtreesOfSize(2)
-
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun subtreesOfSize3() {
-        // Input:
-        //        CONCATENATION
-        //        /  |    |   \
-        //       a   b  UNION  d
-        //               / \
-        //              c   d
-        val child = ParseTreeTest.create1LevelParseTree(ParseTree.UNION, listOf("c", "d"), false)
-        val input = ParseTreeTest.create1LevelParseTree(ParseTree.CONCATENATION, listOf("a", "b", "d"))
-        input.children.add(2, child)
-
-        // Expected:
-        //      CONCATENATION    CONCATENATION
-        //         /   \             /   \
-        //        b   UNION       UNION   d
-        //             / \         / \
-        //            c   d       c   d
-        val expected1 = ParseTreeTest.create1LevelParseTree(ParseTree.CONCATENATION, listOf("b"), true)
-        expected1.children.add(child)
-        val expected2 = ParseTreeTest.create1LevelParseTree(ParseTree.CONCATENATION, listOf("d"), true)
-        expected2.children.add(0, child)
-        val expected = listOf(expected1, expected2)
-
-        val actual = input.subtreesOfSize(3)
-
         assertEquals(expected, actual)
     }
