+++ src/test/java/com/github/giedomak/telepathdb/datamodels/parsetree/ParseTreeUnionPullerTest.kt
@@ -1,8 +1,5 @@
-package com.github.giedomak.telepathdb.datamodels.utilities
+package com.github.giedomak.telepathdb.datamodels.parsetree
 
-import com.github.giedomak.telepathdb.datamodels.parsetree.ParseTree
-import com.github.giedomak.telepathdb.datamodels.ParseTreeTest
-import com.github.giedomak.telepathdb.datamodels.parsetree.ParseTreeUnionPuller
 import org.junit.Test
 import kotlin.test.assertEquals
